delete wallet with incomplete restore

+++ app/src/main/java/com/mw/beam/beamwallet/core/listeners/WalletListener.kt
@@ -124,7 +124,7 @@ object WalletListener {
     @JvmStatic
     fun onImportRecoveryProgress(done: Long, total: Long) {
         LogUtils.log("onImportRecoveryProgress:: done:$done total:$total")
-        returnResult(subOnImportRecoveryProgress, OnSyncProgressData(((done / total) * 100).toInt(), 100), "onImportRecoveryProgress")
+        returnResult(subOnImportRecoveryProgress, OnSyncProgressData(((done.toFloat() / total) * 100).toInt(), 100), "onImportRecoveryProgress")
     }
 
     private fun <T> returnResult(subject: Subject<T>, result: T, responseName: String) {
